+++ kwimage/im_color.py
@@ -273,6 +273,16 @@ class Color(ub.NiceRepr):
         """
         Returns:
             List[str]: names of colors that Color accepts
+
+        Example:
+            >>> import kwimage
+            >>> named_colors = kwimage.Color.named_colors()
+            >>> color_lut = {name: kwimage.Color(name).as01() for name in named_colors}
+            >>> # xdoctest: +REQUIRES(module:kwplot)
+            >>> import kwplot
+            >>> kwplot.autompl()
+            >>> canvas = kwplot.make_legend_img(color_lut)
+            >>> kwplot.imshow(canvas)
         """
         names = sorted(list(BASE_COLORS.keys()) + list(CSS4_COLORS.keys()))
         return names