+++ kwimage/structs/mask.py
@@ -68,7 +68,7 @@ class _Mask_Backends():
         try:
             from pycocotools import _mask
             _funcs['pycoco'] = _mask
-        except ImportError as ex:
+        except Exception as ex:
             if not KWIMAGE_DISABLE_IMPORT_WARNINGS:
                 warnings.warn(
                     'optional module pycocotools is not available: {}'.format(
@@ -78,7 +78,7 @@ class _Mask_Backends():
             try:
                 from kwimage.structs._mask_backend import cython_mask
                 _funcs['kwimage'] = cython_mask
-            except ImportError as ex:
+            except Exception as ex:
                 if not KWIMAGE_DISABLE_IMPORT_WARNINGS:
                     warnings.warn(
                         'optional mask_backend is not available: {}'.format(str(ex)))
@@ -687,7 +687,7 @@ class _MaskDrawMixin(object):
     matplotlib (the ``draw`` method) or opencv (the ``draw_on`` method).
     """
 
-    def draw_on(self, image, color='blue', alpha=0.5,
+    def draw_on(self, image=None, color='blue', alpha=0.5,
                 show_border=False, border_thick=1,
                 border_color='white', copy=False):
         """
@@ -756,6 +756,9 @@ class _MaskDrawMixin(object):
             >>> kwplot.show_if_requested()
         """
         import kwimage
+
+        if image is None:
+            image = np.zeros(self.shape[0:2] + (3,), dtype=np.float32)
 
         dtype_fixer = _generic._consistent_dtype_fixer(image)